Cleanup and move of the Sim class
