+++ src/main/java/com/Resimulators/simukraft/client/render/EntitySimRender.java
@@ -3,13 +3,11 @@ package com.Resimulators.simukraft.client.render;
 import com.Resimulators.simukraft.Reference;
 import com.Resimulators.simukraft.client.data.SkinCacher;
 import com.Resimulators.simukraft.client.model.EntitySimModel;
-import com.Resimulators.simukraft.common.entity.EntitySim;
-import com.google.common.base.Preconditions;
+import com.Resimulators.simukraft.common.entity.sim.EntitySim;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.entity.EntityRendererManager;
 import net.minecraft.client.renderer.entity.LivingRenderer;
-import net.minecraft.client.renderer.entity.layers.ArrowLayer;
 import net.minecraft.client.renderer.entity.layers.BipedArmorLayer;
 import net.minecraft.client.renderer.entity.layers.HeadLayer;
 import net.minecraft.client.renderer.entity.layers.HeldItemLayer;
